chore: replace final dtos in ltp test
ECALC-1813
